Refactoring and improvement of property grid control. Now, it is fully able to interact with keyboard (space bar, arrows).
git-svn-id: http://svn2.assembla.com/svn/secu3man@178 c1dae7de-a428-49f4-85d8-50b65f41d01d

+++ sources/common/MathHelpers.h
@@ -75,7 +75,7 @@ namespace MathHelpers
     for(k = -k2; k <= k2; ++k)
     {
      j = (i + k);
-     if (j >= 0 && j < i_data_size)
+     if (j >= 0 && j < (int)i_data_size)
      {
       op_out[i]+= ip_in[j];
       divisor+= 1;
